major update in code

+++ controllers/Section.js
@@ -1,5 +1,6 @@
 const section = require("../models/Section");
 const course = require("../models/Course");
+const subSection = require("../models/SubSection");
 
 
 exports.createSection = async (res, req) => {
@@ -17,7 +18,7 @@ exports.createSection = async (res, req) => {
         const newSection = await section.create({sectionName});
 
         const updatedCourseDetails = await course.findByIdAndUpdate
-        (courseId , {$push:{courseContent:newSection._id}},{new:true}).populate("section").populate("subSection").exec();
+        (courseId , {$push:{courseContent:newSection._id}},{new:true}).populate({path:"courseContent",populate:{path:"subSection"}}).exec();
 
 
         return res.status(200).json({
@@ -42,7 +43,7 @@ exports.createSection = async (res, req) => {
 exports.updateSection = async (res,req) => {
     try{
 
-        const {sectionName , sectionId} = req.body;
+        const {sectionName , sectionId , courseId} = req.body;
         
         if(!sectionName || !sectionId){
             return res.json({
@@ -54,10 +55,26 @@ exports.updateSection = async (res,req) => {
         const Section = await section.findByIdAndUpdate(sectionId , {sectionName},{new:true});
 
 
-        return res.status(200).json({
-            sucess:true,
-            message:"Section updated sucessfully",
-        })
+        // return res.status(200).json({
+        //     sucess:true,
+        //     message:"Section updated sucessfully",
+        // })
+
+
+        const course = await course.findById(courseId)
+		.populate({
+			path:"courseContent",
+			populate:{
+				path:"subSection",
+			},
+		})
+		.exec();
+
+		res.status(200).json({
+			success: true,
+			message: section,
+			data:course,
+		});
 
 
     }catch(error){
@@ -71,28 +88,76 @@ exports.updateSection = async (res,req) => {
 }
 
 
-exports.deleteSection = async (res , req) => {
-    try{
+// exports.deleteSection = async (res , req) => {
+//     try{
 
-        // get ID - assuming we are getting id by params from frontend
-        const {sectionId} =req.params;
+//         // get ID - assuming we are getting id by params from frontend
+//         const {sectionId} =req.params;
 
-        await section.findByIdAndDelete(sectionId);
+//         await section.findByIdAndDelete(sectionId);
 
 
-        // testing k time check krna hai k mujhe course k schema m se bhi delete krna hai id section ka
+//         // testing k time check krna hai k mujhe course k schema m se bhi delete krna hai id section ka
 
-        return res.status(200).json({
-            sucess:true,
-            message:"Sucessfully deleted section",
-        })
+//         return res.status(200).json({
+//             sucess:true,
+//             message:"Sucessfully deleted section",
+//         })
 
-    }catch(error){
-        console.log(error.message);
-        return res.status(500).json({
-            sucess:false,
-            message:"Something Went Wrong While Deleting Section",
-            error:error.message,
-        });
-    }
-}
+//     }catch(error){
+//         console.log(error.message);
+//         return res.status(500).json({
+//             sucess:false,
+//             message:"Something Went Wrong While Deleting Section",
+//             error:error.message,
+//         });
+//     }
+// }
+
+
+
+exports.deleteSection = async (req, res) => {
+	try {
+
+		const { sectionId, courseId }  = req.body;
+		await course.findByIdAndUpdate(courseId, {
+			$pull: {
+				courseContent: sectionId,
+			}
+		})
+		const section = await section.findById(sectionId);
+		console.log(sectionId, courseId);
+		if(!section) {
+			return res.status(404).json({
+				success:false,
+				message:"Section not Found",
+			})
+		}
+
+		//delete sub section
+		await subSection.deleteMany({_id: {$in: section.subSection}});
+
+		await section.findByIdAndDelete(sectionId);
+
+		//find the updated course and return 
+		const course = await course.findById(courseId).populate({
+			path:"courseContent",
+			populate: {
+				path: "subSection"
+			}
+		})
+		.exec();
+
+		res.status(200).json({
+			success:true,
+			message:"Section deleted",
+			data:course
+		});
+	} catch (error) {
+		console.error("Error deleting section:", error);
+		res.status(500).json({
+			success: false,
+			message: "Internal server error",
+		});
+	}
+};
